+++ src/primitives/triangle/styled.rs
@@ -200,7 +200,6 @@ mod tests {
     #[test]
     fn issue_308_infinite() {
         let mut display: MockDisplay<BinaryColor> = MockDisplay::new();
-        display.set_allow_overdraw(true);
         display.set_allow_out_of_bounds_drawing(true);
 
         Triangle::new(Point::new(10, 10), Point::new(20, 30), Point::new(30, -10))
@@ -212,7 +211,6 @@ mod tests {
     #[test]
     fn it_draws_filled_strokeless_tri() {
         let mut display: MockDisplay<BinaryColor> = MockDisplay::new();
-        display.set_allow_overdraw(true);
 
         Triangle::new(Point::new(2, 2), Point::new(2, 4), Point::new(4, 2))
             .into_styled(PrimitiveStyle::with_fill(BinaryColor::On))
@@ -231,7 +229,6 @@ mod tests {
     #[test]
     fn stroke_fill_colors() {
         let mut display: MockDisplay<Rgb888> = MockDisplay::new();
-        display.set_allow_overdraw(true);
 
         Triangle::new(Point::new(2, 2), Point::new(8, 2), Point::new(2, 8))
             .into_styled(
@@ -260,7 +257,6 @@ mod tests {
     #[test]
     fn off_screen() {
         let mut display: MockDisplay<BinaryColor> = MockDisplay::new();
-        display.set_allow_overdraw(true);
         display.set_allow_out_of_bounds_drawing(true);
 
         Triangle::new(Point::new(5, 5), Point::new(10, 15), Point::new(15, -5))
@@ -348,8 +344,6 @@ mod tests {
 
         let mut display = MockDisplay::new();
 
-        display.set_allow_overdraw(true);
-
         styled.draw(&mut display).unwrap();
         assert_eq!(display.affected_area(), styled.bounding_box());
     }
@@ -382,7 +376,6 @@ mod tests {
         );
 
         let mut display = MockDisplay::new();
-        display.set_allow_overdraw(true);
 
         styled.draw(&mut display).unwrap();